ECER-6368
